refactor: isolate paired phone record type
